edge_device/messenger/requirements.txt
numpy==1.18.4
opencv_python==4.2.0.34
paho_mqtt==1.5.0
